+++ jetty/src/main/scala/org/http4s/server/jetty/JettyBuilder.scala
@@ -31,7 +31,8 @@ sealed class JettyBuilder[F[_]: Effect] private (
     mounts: Vector[Mount[F]],
     private val serviceErrorHandler: ServiceErrorHandler[F],
     banner: immutable.Seq[String]
-)(implicit protected val F: ConcurrentEffect[F]) extends ServletContainer[F]
+)(implicit protected val F: ConcurrentEffect[F])
+    extends ServletContainer[F]
     with ServerBuilder[F]
     with IdleTimeoutSupport[F]
     with SSLKeyStoreSupport[F]